add data to database, config users, test methods

+++ src/main/java/org/example/javafx_objectDB/dao/PeliculaDao.java
@@ -10,5 +10,5 @@ import java.util.List;
 public interface PeliculaDao {
     void guardar(Pelicula pelicula);
     List<Pelicula> listarTodas();
-    int deleteById(Long id);
+    Long deleteById(Long id);
 }
